tree-construction/namespace-sensitivity.dat

#data
<body><table><tr><td><svg><td><foreignObject><span></td>Foo
#errors
#new-errors
(1:1) missing-doctype
#document
| <html>
|   <head>
|   <body>
|     "Foo"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg foreignObject>
|                   <span>
